refactor: move methods from MapdlGrpc class to core and extended
